Split traits from the common crate
