+++ runtime-modules/storage/src/lib.rs
@@ -4,5 +4,6 @@
 pub mod data_directory;
 pub mod data_object_storage_registry;
 pub mod data_object_type_registry;
+pub mod traits;
 
 mod mock;
